yoga_bot/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
diskcache==5.6.3
filelock==3.14.0
fsspec==2024.6.0
huggingface==0.0.1
huggingface-hub==0.23.3
idna==3.7
Jinja2==3.1.4
llama_cpp_python==0.2.77
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
PyYAML==6.0.1
requests==2.32.3
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.1
